+++ packages/contrlibs/ikaros-platform-desktop-client/src/compile/desktop-client-compile.ts
@@ -1,9 +1,0 @@
-import type { CompileServeParame } from '@ikaros-cli/ikaros'
-
-import { ElectronCompileService } from './electron'
-
-export const startDesktopClientCompile = async (
-  parame: CompileServeParame,
-): Promise<void> => {
-  await ElectronCompileService.create(parame)
-}