+++ packages/contrlibs/ikaros-platform-desktop-client/src/compile/electron/index.ts
@@ -1,1 +1,0 @@
-export { ElectronCompileService } from './electron-compile-service'