+++ packages/contrlibs/ikaros-platform-desktop-client/src/compile/electron/main/index.ts
@@ -1,1 +1,0 @@
-export { ElectronMainCompileService } from './electron-main-compile-service'