+++ packages/contrlibs/ikaros-platform-desktop-client/src/compile/electron/preload/index.ts
@@ -1,1 +1,0 @@
-export { ElectronPreloadCompileService } from './electron-preload-compile-service'