+++ packages/contrlibs/ikaros-platform-desktop-client/src/compile/electron/renderer/index.ts
@@ -1,1 +1,0 @@
-export { ElectronRendererCompileService } from './electron-renderer-compile-service'