+++ packages/contrlibs/ikaros-platform-desktop-client/src/index.ts
@@ -1,2 +1,4 @@
+export { ElectronDesktopPlatform } from './platform/electron-desktop-platform'
+export { ElectronDesktopPlatformInstance } from './platform/electron-desktop-platform'
+
 export * from './runner'
-export * from './compile/desktop-client-compile'